app/requirements.txt
fastapi

pydantic
pydantic-extra-types

langchain

python-dotenv
httpx

unidiff



